update the package importing methods
