+++ pyanomaly/config/__init__.py
@@ -1,0 +1,1 @@
+from .defaults import update_config
