--- pyanomaly/datatools/abstract/aug_builder.py
@@ -1,11 +1,0 @@
-# will be deprecated in ver2.0
-class UtilsBuilder(object):
-    def __init__(self, cfg):
-        self.cfg = cfg
-
-    def _build(self):
-        '''
-        build method
-        '''        
-        raise Exception('Not implementations')
-
